hugo

+++ libft-project/Libft-42/ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: ybarbot <ybarbot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/04 17:17:33 by ybarbot           #+#    #+#             */
-/*   Updated: 2023/11/04 17:20:32 by ybarbot          ###   ########.fr       */
+/*   Updated: 2023/11/08 13:44:21 by ybarbot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
 	int	i;
